reworked rule selection to use xml instead of hard coded ui

+++ 3rd_party/oxygine-framework/oxygine/Input.h
@@ -27,8 +27,8 @@ namespace oxygine
         qint32 touchID2index(qint64 id);
         PointerState* getTouchByID(qint64 id);
         PointerState* getPointerMouse();
-        void sendPointerButtonEvent(spStage & stage, MouseButton button, float x, float y, float pressure, qint32 type, PointerState* ps);
-        void sendPointerMotionEvent(spStage & stage, float x, float y, float pressure, PointerState* ps);
+        void sendPointerButtonEvent(spStage & stage, MouseButton button, qint32 x, qint32 y, float pressure, qint32 type, PointerState* ps);
+        void sendPointerMotionEvent(spStage & stage, qint32 x, qint32 y, float pressure, PointerState* ps);
         void sendPointerWheelEvent(spStage & stage, const Vector2& dir, PointerState* ps);
     private:
         explicit Input();
