+++ objects/base/label.h
@@ -47,12 +47,12 @@ public:
      * @brief setWidth
      * @param width
      */
-    virtual void setWidth(float width) override;
+    virtual void setWidth(qint32 width) override;
     /**
      * @brief setHeigth
      * @param height
      */
-    virtual void setHeight(float height) override;
+    virtual void setHeight(qint32 height) override;
 public slots:
     virtual void focusedLost() override {}
 protected:
